Remove es create index

+++ dev_tools/scripts/db_test.sh
@@ -3,6 +3,10 @@
 docker exec db psql -U postgres -d test -c "INSERT INTO local_ega.main(submission_file_path,submission_user,submission_file_extension,status,encryption_method) VALUES('test.c4gh','dummy','c4gh','INIT','CRYPT4GH');"
 
 CONFIGFILE="dev_tools/config.yaml" go run . --action pg_dump
+# bail early since if this fails the rest will fail also
+if [ $? != 0 ]; then
+    exit 1
+fi
 
 docker exec db psql -U postgres -d postgres -c "DROP DATABASE test;"
